Fixed the measurement channels tested on both CS-2 and CS-3
