+++ src/main/java/jcs/controller/cs/SxxBus.java
@@ -1,0 +1,149 @@
+/*
+ * Copyright 2023 Frans Jacobs.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package jcs.controller.cs;
+
+import java.util.Objects;
+
+/**
+ *
+ * @author fransjacobs
+ */
+public class SxxBus {
+
+  private Integer endValue;
+  private Integer max;
+  private String name;
+  private Integer number;
+  private Integer startValue;
+  private Integer type;
+  private Integer length;
+
+  public Integer getEndValue() {
+    return endValue;
+  }
+
+  public void setEndValue(Integer endValue) {
+    this.endValue = endValue;
+  }
+
+  public Integer getMax() {
+    return max;
+  }
+
+  public void setMax(Integer max) {
+    this.max = max;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  public void setName(String name) {
+    this.name = name;
+  }
+
+  public Integer getNumber() {
+    return number;
+  }
+
+  public void setNumber(Integer number) {
+    this.number = number;
+  }
+
+  public Integer getStartValue() {
+    return startValue;
+  }
+
+  public void setStartValue(Integer startValue) {
+    this.startValue = startValue;
+  }
+
+  public Integer getType() {
+    return type;
+  }
+
+  public void setType(Integer type) {
+    this.type = type;
+  }
+
+  public Integer getLength() {
+    return length;
+  }
+
+  public void setLength(Integer length) {
+    this.length = length;
+  }
+
+  public Integer getContactIdOffset() {
+    if (this.number != null) {
+      return (this.number - 1) * 1000;
+    } else {
+      return 0;
+    }
+  }
+
+  @Override
+  public int hashCode() {
+    int hash = 7;
+    hash = 71 * hash + Objects.hashCode(this.endValue);
+    hash = 71 * hash + Objects.hashCode(this.max);
+    hash = 71 * hash + Objects.hashCode(this.name);
+    hash = 71 * hash + Objects.hashCode(this.number);
+    hash = 71 * hash + Objects.hashCode(this.startValue);
+    hash = 71 * hash + Objects.hashCode(this.type);
+    hash = 71 * hash + Objects.hashCode(this.length);
+    return hash;
+  }
+
+  @Override
+  public boolean equals(Object obj) {
+    if (this == obj) {
+      return true;
+    }
+    if (obj == null) {
+      return false;
+    }
+    if (getClass() != obj.getClass()) {
+      return false;
+    }
+    final SxxBus other = (SxxBus) obj;
+    if (!Objects.equals(this.name, other.name)) {
+      return false;
+    }
+    if (!Objects.equals(this.endValue, other.endValue)) {
+      return false;
+    }
+    if (!Objects.equals(this.max, other.max)) {
+      return false;
+    }
+    if (!Objects.equals(this.number, other.number)) {
+      return false;
+    }
+    if (!Objects.equals(this.startValue, other.startValue)) {
+      return false;
+    }
+    if (!Objects.equals(this.type, other.type)) {
+      return false;
+    }
+    return Objects.equals(this.length, other.length);
+  }
+
+  @Override
+  public String toString() {
+    return "SxxBus{" + "name: " + name + ", number: " + number + ", length: " + length + ", contactIdOffset: " + getContactIdOffset() + "}";
+  }
+
+}
