+++ src/main/java/jcs/ui/ControllerPanel.java
@@ -39,7 +39,6 @@ import javax.swing.UIManager;
 import javax.swing.UnsupportedLookAndFeelException;
 import javax.swing.text.BadLocationException;
 import jcs.controller.cs.can.CanMessage;
-import jcs.controller.cs3.devices.LinkSxx;
 import jcs.controller.cs.events.CanMessageEvent;
 import jcs.controller.cs3.events.CanMessageListener;
 import jcs.trackservice.TrackControllerFactory;
